pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ComfyUI_For_Academic"
version = "0.1.0"
description = "Visual algorithm node package based on comfy ui"
authors = [
  {name = "BitWalker", email = "koren.cai.cy@gmail.com"}
]
readme = "README.md"
license = {text = "MIT license"}
classifiers = []
dependencies = [
    "pyserial>=3.5",
    "numpy",
    "opencv-python",
]

[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/BitWalker/ComfyUI_For_Academic/issues"
homepage = "https://github.com/BitWalker/ComfyUI_For_Academic"


[tool.comfy]
PublisherId = "BitWalker"
DisplayName = "ComfyUI_For_Academic"
Icon = ""


[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
